% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfazeD.R
\name{gdistPair}
\alias{gdistPair}
\title{Genetic Distance for a Pair of Samples}
\usage{
gdistPair(
  pair,
  afreq,
  coi,
  nr = 100,
  nm = min(coi),
  rval = NULL,
  reval = NULL,
  equalr = FALSE,
  out = "mle"
)
}
\arguments{
\item{pair}{a list of length two containing two samples.}

\item{afreq}{a list of allele frequencies. Each element of the list
corresponds to a locus.}

\item{coi}{a vector indicating complexity of infection for each sample.}

\item{nr}{an integer value for the resolution of the grid (\eqn{nr - 1}
values between 0 and 1), over which the likelihood will be calculated.
Ignored if non-null \code{reval} is provided.}

\item{nm}{the number of related pairs of strains for evaluation.}

\item{rval}{\eqn{{r}} values for the grid or for evaluation when
\code{equalr} is \code{TRUE}. If \code{NULL}, will be evenly spaced between
0 and 1 and interval \eqn{1/nr}.}

\item{reval}{the grid of \eqn{{r}} combinations, over which the likelihood
will be calculated. A matrix where each column represents a single
combination.}

\item{equalr}{a logical value. If \code{TRUE}, only equal values of \emph{r} for
different pairs of strains are evaluated.}

\item{out}{a character string for the type of results to be returned. If
\code{"mle"}, an estimate is returned, if \code{"llik"} - a vector of
log-likelihood for each combination.}
}
\value{
If \code{out} is \code{"mle"}, a vector of length 1 if \code{equalr}
is \code{TRUE} or length \code{nm} otherwise, containing estimated
\eqn{{r}} (or vector/matrix if not just first). If \code{out} is
\code{"llike"}, a vector of log-likelihood values for each evaluated
combination.
}
\description{
Calculates likelihood over a range of values for relatedness parameter
\eqn{{r}} and provides an estimate for a pair of samples.
}
\seealso{
\code{\link{gdist}} for processing multi-sample data in various
formats and estimating pairwise relatedness.
}
