% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfazeD.R
\name{gdistPair1}
\alias{gdistPair1}
\title{Genetic distance for data with genotyping errors}
\usage{
gdistPair1(pair, afreq, coi, nr = 100, nm = min(coi), rval = NULL,
  reval = NULL, equalr = FALSE, out = "mle", alpha = 0.05)
}
\arguments{
\item{pair}{a list of length two containing two samples.}

\item{afreq}{a list of allele frequencies. Each element of the list
corresponds to a locus.}

\item{coi}{a vector indicating complexity of infection for each sample.}

\item{nr}{an integer value for the resolution of the grid (\eqn{nr - 1}
values between 0 and 1), over which the likelihood will be calculated.
Ignored if non-null \code{reval} is provided.}

\item{nm}{the number of related pairs of strains for evaluation.}

\item{rval}{\eqn{{r}} values for the grid. If \code{NULL}, will be evenly
spaced between 0 and 1 and interval \eqn{1/nr}.}

\item{reval}{the grid of \eqn{\{r\}} combinations, over which the likelihood
will be calculated. A vector if there is a single related pair or if
\code{equalr} is \code{TRUE}; otherwise a matrix where each column
represents a single combination.}

\item{equalr}{a logical value. If \code{TRUE}, only equal values of \emph{r} for
different pairs of strains are evaluated.}

\item{out}{a character string for the type of results to be returned. If
\code{"mle"}, an estimate is returned, if \code{"llik"} - a vector of
log-likelihood for each combination.}

\item{alpha}{significance level.}
}
\value{
\itemize{
\item If \code{out = "mle"}: a vector of length 1 if \code{equalr = TRUE}
or of length \code{nm} otherwise, containing estimated \eqn{{r}} (or vector
/matrix if not just first).
\item If \code{out = "llike"}, a vector of log-likelihood values for each
evaluated combination.
\item If \code{out = "all"}, a list, which contains log-likelihood, MLE,
\eqn{{r}} values corresponding to the acceptance region determined by the
significance level, and the size of that region (as a proportion of all
evaluated).
}
}
\description{
Distance when estimated COI is less than |Ux|: COI increased to |Ux| for a
given locus only. ***Handling of the situation where alleles with estimated
allele frequency of 0 are present: probUxUy crashes only if present in both
Ux and Uy; however, if present in either one, the likelihood is 0, so no need
to calculate.
}
