% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfazeD.R
\name{generateReval}
\alias{generateReval}
\title{Generate a grid of parameter values to evaluate over}
\usage{
generateReval(Meval, rval = NULL, nr = NULL)
}
\arguments{
\item{Meval}{an integer vector.}

\item{rval}{\eqn{{r}} values for the grid.}

\item{nr}{an integer vector of length \code{Mmax} indicating the fineness
of the grid.}
}
\value{
A list of length \code{max(Meval)}. Indices of the elements
correspond to values \code{i} in \code{Meval}; each such element is a
matrix with \code{i} rows. Other elements (if any) are \code{NULL} except
for the first, which is always included.
}
\description{
Generate a grid of parameter values to evaluate over
}
